kconfig: delete zconf.lex.c
This is a file that is automatically generated from zconf.lex.c_shipped
during the kernel compilation process. It has no purpose being kept
around permanently. If it isn't being generated, either kernel sources
are dirty (there are files from a previous compilation process present
directly in the sources) and then there's an attempt to compile the
kernel using an external OUT directory, or some type of fix is needed,
though everything is fine.

Signed-off-by: K9100ii <k9100ii.ki@gmail.com>
Signed-off-by: Aditya Eka Pranata <aepranata@gmail.com>