Revert "Makefile: Fix device not booting with GCC 7.x and above"
This reverts commit 13e9994e25cd687d1057729a40f21d0a406466b8.

-fno-store-merging wouldn't need to be set for the entire kernel. It
could be set for a singular problematic driver.

But, compiling with GCC, I don't seem to get any booting issues/panics
on A40 without this. I initially tried having -fno-store-merging set for
the pl330 driver, which was the cause of panics while not having
-fno-store-merging on kernels I've worked on previously, but then tried
without with no apparent issues.

I guess it can go altogether.

Signed-off-by: K9100ii <k9100ii.ki@gmail.com>